refactor: profile dogprofile mypage 폴더로 이동 , dogprofile dogdetail로 파일명 변경
